(BOLT-664) Upload from StringIO object
This commit adds the ability to upload a StringIO object as a source "file". This allows data from memory to be uploaded to a remote file location via winrm. In the case where the source data is a StringIO object there must only be one object uploaded to a single destination. This is because the StringIO object will not have a file name associated with it. Instead it's "name" is the destination path.